pyproject.toml
[tool.poetry]
name = "python-migration"
version = "0.1.0"
description = ""
authors = ["usiqueira>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
faker = "^25.8.0"
pymongo = "^4.7.3"
psycopg2-binary = "^2.9.9"
tqdm = "^4.66.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
